Initial test for ValidationError
This is an example of how we might implement an generic sort of
ValidationError, where we don't have to explicitly check for errors
(running the validate route with a bad job_id against this code,
for example, will throw a SqlAlchemy error). Presumably we could
update ValidationError to check for specifics, though I'm having
a hard time getting the original exception. Alternately, we could
wrap things like the job id check in a try block (shorter, and
more specific than we use currently). For example: try: do
a db thing except: raise ValidationError('Bad job', 400).